Update views.py
This is the final code for the API requests needed for the feature Profile Management to work. Contains POST, GET, PUT, and PATCH requests. This code allows for the creation of a new user containing the fields username, password, name, email address, and address. Once a username is created it cannot be used again. Given a specific username, the user details are able to be retrieved. A list of the users created can also be retrieved. A user can also update majority of their fields except for the field email. The code also allows for a user to create a credit card for their account needing the cards number, expiration date, and cvv.