Update urls.py
Contains all the urls for the app_name = 'ProfileManagement'. The urls are for the main page, to be able to create a user, being able to retrieve a list of all the users created, being able to retrieve the user details of a specific user, being able to update all the fields of a user except email, being able to update some of the fields of a user except email, being able to create a credit card for a user's account, being able to show the credit card created for a user's account.